Crear el backend para la creación del usuario cliente.
