+++ Backend/StoreFlow.Usuarios/StoreFlow.Usuarios.API/Endpoints/UsuariosEndpoints.cs
@@ -41,32 +41,3 @@ public static class UsuariosEndpoints
     
     
 }
-
-public static class CrearClienteEndpoints
-{
-    public static void MapCrearClienteEndpoints(this IEndpointRouteBuilder app)
-    {
-        app.MapPost("/login/cliente", (CrearClienteRequest crearClienteDto, UsuariosDbContext db) =>
-        {
-            try
-            {
-                crearClienteDto.Validar();
-            }
-            catch (ArgumentException e)
-            {
-                return Results.BadRequest(e.Message);
-            }
-
-            try
-            {
-                db.CrearUsuarioCliente(crearClienteDto);
-                return Results.Created();
-            }
-            catch (UsuarioConCorreoRepetidoException e)
-            {
-                return Results.BadRequest(e.Message);
-            }
-
-        });
-    }
-}
