+++ UI/apps/mobile/src/app/login/login.urls.ts
@@ -2,5 +2,5 @@ import { environment } from '../../environments/environment';
 
 export const LoginUrls = {
   ingresar: `${environment.apiUrl}/usuarios/login`,
-  registrarCliente: `${environment.apiUrl}/cliente`,
+  registrarCliente: `${environment.apiUrl}/usuarios/cliente`,
 };
